treewide: add `vencord-unstable`, add `programs.discord.vencord.unstable`
The main motivation is that Discord's updates sometimes break plugins.
When this happens, you're stuck waiting for several steps:

1. Vencord devs to fix the issue
2. A new release to be published
3. Someone to create a nixpkgs PR
4. The PR to be merged
5. The update to reach nixos-unstable and/or nixos-24.{05,11}

At best, this process takes a few days. Sometimes it can stretch to a
week...

To get around this issue, we're letting users to enable `unstable` so
they can use an "unstable" version that bypasses the release wait time.
We only need to wait for Vencord devs to push a fix to master